pyproject.toml
[tool.poetry]
name = "official-accounts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = {extras = ["speedups"], version = "^3.9.5"}
pycryptodome = "^3.20.0"
python-dotenv = "^1.0.1"
xmltodict = "^0.13.0"


[tool.poetry.group.dev.dependencies]
aiohttp-debugtoolbar = "^0.6.1"
aiohttp-devtools = "^1.1.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
